Navigation Component #1
Made bottom navigation bar to work with navigation component (nav graph, nav controller, etc)

If you run the program as is, the bottom nav bar should work for budget and accounts

followed the tutorial here to do this:
https://medium.com/@freedom.chuks7/how-to-use-jet-pack-components-bottomnavigationview-with-navigation-ui-19fb120e3fb9
tutorial here gives some background knowledge:
https://developer.android.com/codelabs/basic-android-kotlin-training-fragments-navigation-component?authuser=3&continue=https%3A%2F%2Fdeveloper.android.com%2Fcourses%2Fpathways%2Fandroid-basics-kotlin-unit-3-pathway-2%3Fauthuser%3D3%23codelab-https%3A%2F%2Fdeveloper.android.com%2Fcodelabs%2Fbasic-android-kotlin-training-fragments-navigation-component#6

added a placeholder for Transactions since it might not use a normal fragment, but a dialog fragment. Leaving the decision to the one managing that part.